+++ src/expressionEvaluator/context.ts
@@ -1,0 +1,1 @@
+export interface ExpressionContext {}
