+++ src/expressionEvaluator/evaluator.ts
@@ -1,0 +1,5 @@
+import { ExpressionNode } from './node'
+
+export interface ExpressionEvaluator {
+  evaluate(expression: ExpressionNode, context: any): any
+}
