--- src/expressionEvaluator/expressionContext.ts
@@ -1,1 +1,0 @@
-export abstract class ExpressionContext {}
