--- src/expressionEvaluator/iExpressionEvaluator.ts
@@ -1,5 +1,0 @@
-import { JSExpr } from './jsExpr'
-
-export interface IExpressionEvaluator {
-  evaluate(expression: JSExpr, context: any): any
-}
