+++ src/expressionEvaluator/index.ts
@@ -1,1 +1,10 @@
-export { ExpressionEvaluator } from './expressionEvaluator'
+export type { ExpressionContext } from './context'
+
+export type { ExpressionEvaluator } from './evaluator'
+
+export type { ExpressionFunction } from './function'
+
+export type { ExpressionNode, ExpressionNodeEvaluator } from './node'
+export { ExpressionNodeType } from './node'
+
+export { JavascriptExpressionEvaluator } from './javascript/evaluator'
