--- src/expressionEvaluator/jsExpr.ts
@@ -1,5 +1,0 @@
-import { ExpressionNodeType } from './expressionNodeType'
-
-export interface JSExpr {
-  type: ExpressionNodeType
-}
