+++ src/expressionEvaluator/node.ts
@@ -1,3 +1,6 @@
+import { ExpressionEvaluator } from './evaluator'
+import { ExpressionContext } from './context'
+
 export const enum ExpressionNodeType {
   BinaryExpression = 'BinaryExpression',
   CallExpression = 'CallExpression',
@@ -9,4 +12,13 @@ export const enum ExpressionNodeType {
   MemberExpression = 'MemberExpression',
   ThisExpression = 'ThisExpression',
   UnaryExpression = 'UnaryExpression',
+}
+
+export interface ExpressionNode {
+  type: ExpressionNodeType
+}
+
+export interface ExpressionNodeEvaluator {
+  evaluator: ExpressionEvaluator
+  evaluate(expressionNode: ExpressionNode, context: ExpressionContext): any
 }
